Move creation of builtin readers to the participant factory
